[CHANGE] Switch fleet comms to the new autocomplete library
