+++ fleetpings/templates/fleetpings/index.html
@@ -68,12 +68,11 @@
 {% block extra_javascript %}
     {% include "bundles/jquery-datetimepicker-js.html" %}
     {% include "bundles/clipboard-js.html" %}
-    {% include "fleetpings/bundles/flexdatalist-js.html" %}
+    {% include "fleetpings/bundles/bootstrap5-autocomplete-js.html" %}
     {% include "fleetpings/bundles/fleetpings-js.html" %}
 {% endblock %}
 
 {% block extra_css %}
     {% include "bundles/jquery-datetimepicker-css.html" %}
-    {% include "fleetpings/bundles/flexdatalist-css.html" %}
     {% include "fleetpings/bundles/fleetpings-css.html" %}
 {% endblock %}
