+++ fleetpings/templates/fleetpings/partials/form/segments/fleet-comms.html
@@ -1,5 +1,5 @@
 {% if fleet_comms|length > 0 %}
-    <datalist id="fleetCommsList">
+    <datalist id="fleet-comms-list">
         {% for fleet_comm in fleet_comms %}
             <option value="{{ fleet_comm }}">{{ fleet_comm }}</option>
         {% endfor %}
